moved the files into modificationTests folder
