+++ hedera-mirror-test/k6/src/web3/test/index.js
@@ -62,16 +62,16 @@ import * as contractCallEstimateUnfreezeToken from './contractCallEstimateUnfree
 import * as contractCallEstimateUnfreezeNft from './contractCallEstimateUnfreezeNft.js';
 import * as contractCallEstimateTransferToken from './contractCallEstimateTransferToken.js';
 import * as contractCallEstimateTransferNft from './contractCallEstimateTransferNft.js';
-import * as contractCallPrecompileApprove from './contractCallPrecompileApprove.js';
-import * as contractCallPrecompileAssociate from './contractCallPrecompileAssociate.js';
-import * as contractCallPrecompileCryptoTransferHbars from './contractCallPrecompileCryptoTransferHbars.js';
-import * as contractCallPrecompileCryptoTransferToken from './contractCallPrecompileCryptoTransferToken.js';
-import * as contractCallPrecompileDeleteToken from './contractCallPrecompileDeleteToken.js';
-import * as contractCallPrecompileDissociate from './contractCallPrecompileDissociate.js';
-import * as contractCallPrecompileMintToken from './contractCallPrecompileMintToken.js';
-import * as contractCallPrecompileNestedAssociate from './contractCallPrecompileNestedAssociate.js';
-import * as contractCallPrecompileTransferFungibleToken from './contractCallPrecompileTransferFungibleToken.js';
-import * as contractCallRedirectApprove from './contractCallRedirectApprove.js';
+import * as contractCallPrecompileApprove from './modificationTests/contractCallPrecompileApprove.js';
+import * as contractCallPrecompileAssociate from './modificationTests/contractCallPrecompileAssociate.js';
+import * as contractCallPrecompileCryptoTransferHbars from './modificationTests/contractCallPrecompileCryptoTransferHbars.js';
+import * as contractCallPrecompileCryptoTransferToken from './modificationTests/contractCallPrecompileCryptoTransferToken.js';
+import * as contractCallPrecompileDeleteToken from './modificationTests/contractCallPrecompileDeleteToken.js';
+import * as contractCallPrecompileDissociate from './modificationTests/contractCallPrecompileDissociate.js';
+import * as contractCallPrecompileMintToken from './modificationTests/contractCallPrecompileMintToken.js';
+import * as contractCallPrecompileNestedAssociate from './modificationTests/contractCallPrecompileNestedAssociate.js';
+import * as contractCallPrecompileTransferFungibleToken from './modificationTests/contractCallPrecompileTransferFungibleToken.js';
+import * as contractCallRedirectApprove from './modificationTests/contractCallRedirectApprove.js';
 import * as rampUp from './rampUp.js';
 
 // add test modules here
